feat(mobile): add basic structure and login
Basic structure: Theme provider, navigator and toast provider.
Login: Composed of two screens, <LandingSreen /> and <LoginScreen />. The landing asks for the user the domain of his account
if the user provides a wrong domain it will toast him that there's no domain with that name, otherwise will prompt success and
move him to the login page; Whereas the domain is already known what remains of needed information to accomplish the login is
the user email and password, the form provide this and try to log, the toast mechanics are the same except the api this time
returns in the header a set-cookie directive which shall append to the local cookie of the page the AccessToken and RefreshToken
used.

NOTE: I'm not sure if RN/Expo deals with that plain simple, I managed to find a cookie manager under react-native-cookies/cookies
with last commit in 5 months or so, thusly it sounds iffy at best, yet still I have not confirmed if the cookies were kept or not.

Signed-off-by: Daniel Boll <danielboll.academico@gmail.com>